feat(foundation): implement zero-drift foundation & legacy quarantine
